+++ packages/storefront-cart/src/lib/fetchers.ts
@@ -3,7 +3,7 @@ import convertCartToInput from "./utils/convertCartToInput";
 import { createCartMutation } from "./queries";
 
 export async function handleNoRESTCart() {
-  const res = await client.query(createCartMutation);
+  const res = await client.query({ query: createCartMutation });
   if (!res?.data?.cartCreate?.cart?.id)
     throw new Error("Cart creation failed!");
   return res.data.cartCreate.cart.id;
@@ -18,7 +18,8 @@ export async function handleHasRESTCart() {
   })
     .then((res) => res.json())
     .then((cart) => convertCartToInput(cart));
-  const res = await client.query(createCartMutation, {
+  const res = await client.query({
+    query: createCartMutation,
     variables: {
       input: cart,
     },