+++ packages/storefront-client/src/lib/constants.ts
@@ -1,0 +1,5 @@
+export const CLIENT = "Bryt Storefront Client";
+export const NO_DATA_OR_ERRORS_ERROR =
+  "An unknown error has occurred. The API did not return a data object or any errors in its response.";
+export const GQL_API_ERROR =
+  "An error occurred while fetching from the API. Review 'graphQLErrors' for details.";